Updates to employment task
New designs consolidate the separate employment tasks into a single
task. In order to preserve the tasks completed on previous years claims
we add the other employment tasks to the tasks array only if they are
persisted. The employment task on old claims will now show some extra
information it's not worth the complexity to avoid this.

We likely want to rethink the task model.
Currently tasks are only persisted when they are completed ("passed",
"failed"), switching to an approach where when the claim is submitted we
persist the list of tasks that need completing would allow us to change
the tasks list on future claims without effecting the task list of
claims in previous years.