Switch match model for array of ids
Remove the `Claims::Match` model and store the previous year's claims as
an array of ids on the eligibility.
IRP isn't going to be around for much longer so we want to keep the
blast radius of any new features as small as possible. The
`Claims::Match` model has the potential to be re(mis)used in other
scenarios with potentially unintentional side effects on old IRP claims.
Usually storing ids like this in an array is a smell but here it seems
to work quite well.