Add create employment history
Adds the ability for admins to record an employment history.
This is a bit of a wip commit as we need to support deletion and add a
school search.
We're storing the employment history as a json column rather than a
separate model to keep the size of this feature smaller, as only IRP
needs this model and IRP will be removed next year.