Add autocomplete to school field
The design on the form calls for a school autocomplete field. This
existing school autocomplete js files won't work with this form's
design as they take over the form, assuming the school search is the
single input on the form, as this form has other inputs along with the
school search we've implemented a new school search that only takes over
a single field. We could replace the existing school searches with this
approach, but that's something for a separate ticket.
We'll also need to come up with a non JS fallback for this UI, again to
be addressed in a separate ticket.