Record deletions
We've implementing soft delete to act as an audit trail for who's
removed employments. We may want to consider wrapping the employment
history array with some method to automatically hide deleted
employments, however as we're only using this for the one task we can
get away with checking `deleted?` in the view.