Remove redirect_to_first_slug
We want to keep all navigation logic in the navigator rather than
referencing the slug sequence directly in the controller.

To support this we need to update the EY policy eligibility checker to
handle the email not yet being provided.

Have deleted the form submissable test, fixing the test would require
implementing a journey and using the real navigator (test's `next_slug`
just pulls the next slug from the array). Hopefully we'll be able to
remove all this code soon!