Remove `redirect_to_slug`
This is only called by `next_slug`. We can hardcode the claim_path to
make things easier to read as this code is only ever used in the claims
controller (we'll be removing this concern and inlining it into the
controller).
We can remove the guard clauses too as only one page still uses the call
backs, and we're working on removing that.