Remove form_submittable_spec
This spec is really hard to maintain as we remove functionality from the
form submitable module. The spec defines it's own controller and routes
and requires a lot of setup to mimic what's in the claims controller.
All the functionality of this module is covered by the journey feature
specs already.