Migrate still teaching form to use form builder
Previously the still teaching form and associated views had multiple
branches based on whether the school was a recent tps school or selected
during the claim. In order to make migrating to form builder easier
we've extracted a separate form for the recent tps school branch and
moved the branching up into the slug sequence.