Handle extended attributes on eligibility
Bit of a tricky one and not super stoked on the implementation.
We want to find all relevant claims that have non null pii attributes
and nullify these attributes. The attributes may be on the eligibility
so we have to join to that.

We could consider having separate constants for the claim and
eligibility attributes, however I think a single constant is a bit more
readable, even if we have to add logic to the scrubber to pluck out the
appropriate attributes.

Another option was to use an timestamp to track extended attribute
scrubbing, which would remove the queries checking if the extended
attributes were null (and the need to join to eligibility) how ever this
would require passing in the timestamp column name to the claim
scrubber, which felt weird.

The generated sql doesn't come out _too_ bad

```sql
SELECT "international_relocation_payments_eligibilities".*
FROM "international_relocation_payments_eligibilities"
INNER JOIN "claims"
  ON "claims"."eligibility_type" = 'Policies::InternationalRelocationPayments::Eligibility'
  AND "claims"."eligibility_id" = "international_relocation_payments_eligibilities"."id"
WHERE (
  "claims"."id" IN (
    SELECT "claims"."id"
    FROM "claims"
    INNER JOIN "decisions" ON "decisions"."claim_id" = "claims"."id"
    WHERE "claims"."policy" = 'Policies::InternationalRelocationPayments'
    AND "decisions"."undone" = FALSE
    AND "decisions"."approved" = FALSE AND (
	    decisions.created_at < '2022-08-31 23:00:00'
    )
  )
  OR "claims"."id" IN (
    SELECT "claims"."id"
    FROM "claims"
    INNER JOIN "decisions" ON "decisions"."claim_id" = "claims"."id"
    INNER JOIN "claim_payments" ON "claim_payments"."claim_id" = "claims"."id"
    INNER JOIN "payments" ON "payments"."id" = "claim_payments"."payment_id"
    INNER JOIN "payroll_runs" ON "payroll_runs"."id" = "payments"."payroll_run_id"
    WHERE "claims"."policy" = 'Policies::InternationalRelocationPayments'
    AND "decisions"."undone" = FALSE
    AND "decisions"."approved" = TRUE
    AND 1=1
    AND (payments.scheduled_payment_date < '2022-08-31 23:00:00')
  )
)
AND "international_relocation_payments_eligibilities"."passport_number" IS NOT NULL
AND (
  "claims"."national_insurance_number" IS NOT NULL
  OR "claims"."first_name" IS NOT NULL
  OR "claims"."middle_name" IS NOT NULL
  OR "claims"."surname" IS NOT NULL
)
```