Add passport number to retained pii
Passport number is stored on the eligibility, so we can no longer use a
simple where clause to find claims with retained pii attributes (we have
to join to the eligibility and filter that). To find the claims with
retained pii attributes we now use the
`retained_personal_data_removed_at` timestamp.