Add current_organisation_ukprn to DfeSignIn::User
In order to support the FE provider dashboard we need to be able to
return claims a provider has access to. Schools in our DB don't have a
`dfe_sign_in_organisation_id`, however they do have a `ukprn` which DSI
returns. In order to use this we need to update the auth session to
capture this and set it on the user, which we do in a future commit.
We've named this column `current_organisation_ukprn` as when the user
comes back to use from DSI after switching organisations we'll update
this column with the ukprn for the organisation they're signed into.
We've added a struct `Organisation` to wrap this column to support
hanging convenience methods off.