further_education_payments auth skeleton
Wires up the basic controllers required for viewing the claim dashboard
with place holder copy and unmade decisions.
There's some feature specs written that test all this but we'll commit
these as part of the actual role and experience wizard feature.

As per the existing FE provider journey we support DSI bypass in
development.

There's a bunch of clean up to do before this can be merged!