Add the initial role and experience form
Adds the role and experience form, without branch on contract type and
check answers screen.
We can't reuse the navigator for managing the wizard state as it cares
about eligibility and other attributes of the claimant journey. The
wizard for verifying FE claims is simple enough that we can just return
the first incomplete form.
This commit doesn't include save and continue functionality, nor does it
have the declaration part of the check answers form, so check answers is
the furthest the verifier can complete.

Each form in the wizard will implent a form object, a view template, and
attributes to the eligibility. There's a bunch of form behaviour that
could be shared by a base form but we'll worry about that and more
complex branching logic when we have to.