Extract wizard management
The conditionals in the journey are getting a little more complex so we
want to extract something to manage what form to render next. The
`Wizard` is like a combination of slug sequence and navigator. It
decides what forms are part of the journey based on the eligibility and
then returns the first incomplete form.
It also handles clearing answers set on the eligibility from forms that
are now no longer reachable (eg when the user goes down a new branch
after chaning an answer), though this functionality isn't strictly
needed until we implement the change links on the check answers page.