Handle changing answers
The design we need to support means we have to complicate the wizard
navigation logic.

Depending on the contract type option chosen we may render an additional
screen after role and experience. The change link on the check answers
page only goes to the first role and experience screen, so when dealing
with a non perm contract we need to make sure we show the second screen
even though that form is complete.