Add save and return later
We allow the verifier to partially submit the form if they choose false
for the section complete option, this allows them to go off an check
some information without losing progress.
To support this we conditionally allow nil for any values on the form
(other than the section complete answer). We still however want to
ensure that any answers provided fall into the expected list of answers.
Rails validators don't accept a callable for `allow_nil`. Passing a
symbol results in always allowing `nil` as it's truthy. We could
duplicate the validations, and use if and unless, or conditionally
append nil to the list of options we're validating against however both
of these approaches are really noisy. Instead we've implemented our own
validator "Included" which works the same as the built in "Inclusion"
validator but accepts a callable option for allow nil.
**Note** there is one important issue here, passing a second validation
to the same validates call will always allow `nil` values, eg

```
validates(
  :provider_verification_teaching_responsibilities,
  included: {
    in: ->(form) { form.teaching_responsibilities_options.map(&:id) }
  },
  format: {
    with: /SomeRegex/
  },
  allow_nil: :save_and_exit?
)
```
`nil` will always pass the format check as the built in format validator
will see `:save_and_exit?` as truthy rather than attempt to call it
(will send a pr to rails to address this).

Other note is the normalisation callback we've introduced in the base
form. First we can't use the normalises macro without including a bunch
of modules to mimic active record. Second we don't want to use
`presence` here as it will convert `false` to `nil` so we need the
explicit empty string check.