Adds the verification declaration
Upon completing the wizard we store the verifier and set the field
`provider_verification_completed_at` timestamp. We've agreed that a non
null value for `provider_verification_completed_at` represents a claim
as having been verified.

Note we use a presence validation, rather than an acceptance validation,
for the declaration checkbox as the acceptance validation doesn't run if
the attribute is `nil` so the wizard always thinks this form is
complete.