Handle attempted edits of verified claims
If a claim has been verified we don't want to allow verifiers to modify
answers. We need somewhere to redirect them to, as such we've
implemented an initial version of the read only claim view. We've put
this in verification/show for the time being, could also go in
verified_claims/show