Add employment checks
If the claimant has failed one login IDV then we require the provider to
verify their identity. To do this we conditionally add the `claimant_*`
forms to the wizard (the "employment check" section of the wizard.

There are now two exits from the wizard:
1.) The provider completes the whole wizard as before
2.) The provider indicates the claimant isn't employed by the college
    and exists the wizard early

In the second scenario we still want to mark the claim as verified,
which we do by setting the same fields on the eligibility as in the
"standard" wizard journey (
`provider_verification_completed_at`,
and `provider_verification_verified_by_id`
)

Claims where the provider indicates the claimant isn't employed by the
college are considered by the business to have been rejected by the
provider.

There's also some complexity around flash messages. There's 3 scenarios
where we show a flash message
* Wizard is completed like normal
* Wizard is exited early
* Employment checks section is complete, but the verification forms are
  still to do.

We've moved determining which message to show into the wizard as it's
dependent on the wizard's state, however I'm not super stoked on this,
maybe a better approach is to use I18n and pass a wizard state key
¯\_(ツ)_/¯

There's also a new verified claims show variant for showing provider
rejected claims.

One of the requirements was to show the same validation errors on the
claimant personal details form as the personal details form used by
other services, as such we have to get tricky with out multipart date
param handling in order to support showing errors for a missing day etc.
We should extract a concern `DateOfBirth` which handles all the
validation and param noise and share that with the personal details
form. We decided not to implement this at the minute as it requires
changing the personal details html and there's already enough changes in
this commit!