Extract DateOfBirth concern
We have some rules around validation errors we show to users that are
tricky to implement when we parse multipart date params using
ActiveRecord::AttributeAssignment.
We will want to share this date of birth validation handling with other
forms so we're extracting a concern.
To avoid having to overwrite the initializer in the module (other
forms may need to reopen the initializer) we've introduce active model
callbacks allowing us to define an after initializer callback the DoB
handling can hook into.

Customising the error messages is done by overwriting methods in the
including class. We've taken this approach rather than dealing with i18n
as some consumers of the module will need to pass variables to i18n (eg
`claimant_name`) which is a pain to handle in the module!