Add EY provider alt idv journey
If the claimant is unable to prove their identity on one login we'll
send an email to the provider to verify their identity ("the alt idv
journey"). To handle this we've implemented a new journey.

Couple of design decisions to call out
* Finding the claim
We're going to send an email to the provider containing a magic link for
them to verify the claim. We're currently using the claim's reference
number for this. The alt idv journey shows information about the
claimant so we may want to consider changing to a longer code to find
the claim.

* "Ineligible claim"
If the provider says that the claimant does not work at the nursery the
claimant is ineligible and we show a dead end page. We've chosen to use
the slug sequence for this branching rather than the eligibility check
as the journeyy the provider is on is not ineligible. When we introduce
the handling for providers attempting to verify an already verified
claim it would make sense to use the eligibility checker for this.