Copy answers to claim
We may exit the verification journey early if the provider says the
claimant doesn't work at the nursery.
The time stamp `alternative_idv_completed_at` indicates that alternative
idv has been completed on a claim. We can later compare the claimant and
provider answers to determine if alternative idv is a pass or a fail.