Handle revisits to the verification link
We need to deal with the provider visiting the link for already verified
claims. We're treating this as "ineligible". To avoid the
ineligibility checker kicking providers out as soon as a claim has been
verified we check if the claim was verified in the current session.
We also handle the edge case where the claim reference param is dropped
for some reason.