Notify subscribers when alternative idv completed
Forms are conditionally added to the wizard when alternative idv is
required. There's two exits from the wizard if alternative idv is
required. The first exit is if the claimant is not employed by the
college. The second exit is when all wizard steps have been completed.
In both cases we want to trigger the alternative idv completed business
rules. The rules will handle what to do with claims that have passed /
failed idv.