Add the ability to resend verificaiton emails
Easier to not reuse the shared verification email partial here. We don't
have another screen to send the user to, like in the usual email
verification journey, as we're not sending an email to an address
provided by the user, so we reuse the form and branch on the presence of
the "resend" param.