Show if claim already verified on link visit
Rather than having the provider click the start now button to see if a
claim has already been verified we want to show that a claim has already
been verified on page load.
Bit ugly writing a query directly in the view but it seems a better
option than making the landing page controller know about this specific
journey, or moving the landing page inside the journey such that we need
to create a journey session with a get request.