Update EY alternative verification task
We now want to split the bank details and personal details sections of
this task and remove the automated task passing. Currently the task
verifier still runs the automated task passing but we'll remove that in
a future commit as I suspect we'll want to reintroduce automated task
passing.

We've introduced `Claim#hmrc_name_match` which returns the name match
column from the last hmrc response. We picked the last response as this
is the response from when the bank details form is submitted.