Auto pass the personal details sub task
Brings back some of Phil's code for passing the ey task but modifies it
to handle passing the sub tasks.

This task is a bit weird in that there are two "sub tasks" associated.
Usually a task is considered completed when it's persisted, however the
ey_alternative_verification_form can be persisted without being
completed.

If the ey_alternative_verification verifier passes the personal details
section of the task, indidicated by the
`personal_details_were_passed_automatically` attribute, then we don't
show the personal details radio button in the form.

We likely need to revisit how `Task`s work, and introduce a proper
notion of subtasks.