Display the ey_alt verification task if it exists
We have a claim that has completed the Y1 provider journey but failed
alt idv. We want to send the claimant a link to complete the
practitioner journey however as the claim is a Y1 claim the alt idv task
wont show. When the provider has completed the alt idv the background
job we kick off will create the task, see
`Policies::EarlyYearsPayments#alternative_idv_completed`