Update query to work with year 2 fe claims
Year 2 claims have a different provider verification mechanisim so, in
order to get the admin filters to work as expected, we need to update
this query to work with them. We also need the filters to support year 1
claims.

We need to do some clean up on the feature specs as a bunch of them are
still working with year 1 claims and this is becoming a bit of a pain to
maintain.