Fix EY claim identity task
For some EY year 1 claims the practitioner journey was completed after
we removed the Identity claim verifier from EY. There for the identity
task wasn't passed, nor was `claim_verifier_match` set on the task, this
caused the identity task on these claims to be rendered in as incomplete
and the form to complete them was not shown. These claims did have the
one_login_identity (OLIdentity verifier replaced Identity verifier) so
we want to render the `one_login_identity` task rather than the
`identity_confirmation` task if it's present.