Update admin tasks to show FE qualification reason
Updates the provider verification task to show the reason for not
enrolling in a teaching qualification if applicable. The requirement is
to show either "Valid reason" or "No valid reason" rather than the
actually option selected. The actual value the provider selected is
available on the view full claim page (backed by the
eligibility_admin_answers_presenter)