Rescue missing submitted claim
If the session has been cleared and the claimant attempts to visit the
confirmation page we we're throwing an error. This commit changes the
behaviour such that we send them back to the journey start page.

We may want to consider doing something else, such as putting the claim
reference in the url on submission and using that to find the submitted
claim. However as the confirmation page shows some details about the
claimant we've opted against that for the time being.