Move slug sequence use into navigator
Slug sequence is very closely tied to the navigator. Rather than passing
in an instantiated slug sequence and fetching the journey session from
that, pass in the journey session and build the slug sequence in the
navigator.
Also navigator is now responsible for checking for dead end slugs when
determining if there is a previous slug.