Move slug sequence use into navigator
Slug sequence is very closely tied to the navigator. Rather than passing
in an instantiated slug sequence and fetching the journey session from
that, pass in the journey session and build the slug sequence in the
navigator.
Also navigator is now responsible for checking for dead end slugs when
determining if there is a previous slug.

Not happy about having to introduce the struct to trick the navigator
into thinking there's a journey session to pull the journey from (the
only bit of info it needs to render the completed page)