Fix resend email link
If a user visits an expired verification link in a different browser (or
after their session is purged) the resend link throws a 500 error.
The error occurs as there is no journey session for EY PV Start journey,
so when the navigator attempts to check if the "email-address" slug is
in sequence it calling methods on `nil`.
The core of the issue is the "Resend email" link takes the user to a
different journey (EY PV Start) from the expired link page (EY PV
Authenticated). These journeys have different routing names and so
create different journey sessions.
To fix this issue we rely on how the claim's controller handles the skip
landing page param. If the skip landing page param is present the claims
controller will create a new journey session if one doesn't exist.

We may want to consider combining the start and authenticated journeys
into a single provider journey.