Record steps on session
When debugging errors it's useful to see the path through the wizard the
claimant has taken. We've added a new column to journey sessions to
track the steps completed, and we've updated our monitoring to include
this information.

We may want to consider introducing a report showing the number of
sessions stopped at each step to highlight any service issues.