Remove redundant ECP-related eligibility check
Previously the schools Targeted Retention Incentive Payments journey also supported a second policy - Early Career Payments. In this prior version of the journey, an additional ITT subject question radio option was present for foreign languages, as this subject was acceptable for that policy. Since the ECP journey was removed, this radio option will no longer be displayed, so we no longer need this check.

Additionally, the return value `subject_invalid_for_tslr` is incorrect in this context, since TSLR refers to the Teacher Student Loan Repayments policy, which is supported by a separate journey. I assume this was a typo/confusion of acronyms.