Rename provider verification strings
We want the strings that back the provider options to match the strings
that back the claimant options to make comparison easier.
There's no entries in `provider_verification_teaching_hours_per_week` on
production but we've added the migration to not break any review apps
that are still in use.