Handle edge case
The provider and claimant answers can differ for the full ay contract
question. If the claimant has said yes to the full ay contract but the
provider has said no, and thus had to answer the taught at least one
term question, we need to show the taught at least one term question on
the admin task.