Always show N/A
We always want to show "N/A" for the claimant answers to "Contract of
employment", "Taught at least one academic term", and "Full academic
year" even if the claimant has provided a response.
Reasoning from jira ticket CAPT-2793

>  I think the concern from policy was that it is the Provider’s answer
only that should be dictating whether the verification task fails for
these 3 questions. The claimant’s response is irrelevant. Therefore the
thinking from policy is that having N/A there should make it easier for
ops to identify the reason for failure.