Move claim methods onto the provider
Adds claim methods to the eligible fe provider.
We're going to want to reuse these methods in the summary emails we send
to the provider, so moving the calculations out of the controller on to
the provider to keep everything in one place.

We've introduced a new method `current_provider` to the controller.
We'll want to replace the `claim_scope` method with accessing the claims
on the provider model.
We've used methods rather than associations on the provider as `Claim`
is used between multiple policies. This is not very composable, instead
we could do something like:
* define a has one association between fe_provider and school
* define an association between eligibility and school
* define an association between claim and fe eligibility
* extend the claim association to include our FE specific scopes
this seems a bit overkill for our current needs, though it's worth
keeping in mind if defining more and more methods gets painful.