Refactor to scopes
Really not sure if this mad or not.
Adds scopes to EligibleFeProvider for working with claims.
As claims are shared between various policies we don't want to polute
the claim model with these policy specific scopes, instead we extend the
claim association on EligibleFeProvider to add the scopes the provider
requires. Also we have to do quite a bit of faffing to reach claims from
the provider. Not sure how much I like this, but it is nicer to use than
the previous method based approach.