Early years providers search
The ops team currently maintain a spreadsheet of claims submitted by
provider. We have this information in the system. This commit adds a
simple searchable dashboard for ops to check the number of claims
submitted by a given provider against their claim allowance.
Theres ~450 providers on production. We're only returning a
handful of attributes from the DB, and this page isn't likely to be
frequently visited so we can trade off pagination for ability to scan
the whole data set.