Don't hide personal data we still have
There's two tiers of personal data scrubbing. Some data we scrub at the
end of the academic year, some data we scrub after 5 years (varies by
policy). We can't solely rely on the `personal_data_removed?` flag to
determine if we should show the personal data removed method, so we
check that and check if the attribute value is blank.