Adds early years teachers provider
Adds the initial scaffolding for the early years teachers provider
journey. The plan is to build the prototype for this service in this
application and issue a link to a review app for user testing.
As we're building a prototype we'll likely not need to implement
eligibility checking and just minimal slug branching, as such we've not
added the EarlyYearsTeachers policy code, only the journey.
We've added a place holder initial question "favourite colour" to make
sure everything is wired up. This will need to be removed once we start
work on the initial prototype screens.