Admin tasks list overview
This page introduces the admin task list overview, allowing admins to
view task statuses for in progress claims.
We also provide filters for admins to find claims with a specific task
status.
Task statuses are currently calculated in memory, so we can't use
pagination. Testing with several thousand claims, the page takes about 5
seconds to load (once we disable the matching claim check task), with
most of that time spent in the view layer. As we're only showing
inprogress claims it's very unlikely we'll hit these numbers on
production!